+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Upcoming release
 * fix fetching device dnd for fresh installations
+* fix custom params & headers on cloud API requests
 
 ## 2.2.4
 * check partially fetched notifications when fetching inbox items